requirements.txt
torch
torchvision
wandb
python-dotenv
scipy
matplotlib
tqdm
gurobipy
pandas
plotly
seaborn
kaleido